Rebranding a Mercalin, licencias v2 firmadas (prueba de 7 días), etiquetas por peso con código de barras único, y mejoras de UX en Dashboard/Productos/Etiquetas
- Licencias: nuevo formato autocontenido HMAC-firmado (payload+vencimiento), reemplaza el hash simple; habilita período de prueba de 7 días con pantalla de vencimiento (TrialExpired.tsx)
- Etiquetas por peso: al imprimir una etiqueta para un producto pesable se genera un código de barras único (tabla weighed_labels) que Caja reconoce al escanear y cobra el peso/precio exacto grabado en esa etiqueta
- Productos: unifica los botones de Stock y Recibir mercadería en un solo flujo; corrige búsqueda rota por columna ambigua tras el JOIN con proveedores; etiqueta los campos de precio/costo "por kg" en productos pesables
- Dashboard: tarjetas de Alertas y Consejos del día resumidas con modal "ver todas"; campanita de notificaciones en el sidebar y en modo foco
- Recordatorio (no bloqueante) de cerrar caja al intentar cerrar la ventana con una sesión abierta
- Etiquetas: nombres largos hacen wrap en vez de cortarse
- Rebranding completo a Mercalin: ícono, nombre de producto/ventana, instalador

+++ src-tauri/src/commands/returns.rs
@@ -54,13 +54,13 @@ pub fn create_return(input: NewReturn, state: State<AppState>) -> CmdResult<Retu
         .map_err(err)?;
 
         if let Some(pid) = item.product_id {
-            let qty_before: i64 = tx
+            let qty_before: f64 = tx
                 .query_row("SELECT stock FROM products WHERE id=?1", params![pid], |r| r.get(0))
-                .unwrap_or(0);
+                .unwrap_or(0.0);
 
             tx.execute(
                 "UPDATE products SET stock=stock+?1, updated_at=CURRENT_TIMESTAMP WHERE id=?2",
-                params![item.qty as i64, pid],
+                params![item.qty, pid],
             )
             .map_err(err)?;
 
@@ -69,9 +69,9 @@ pub fn create_return(input: NewReturn, state: State<AppState>) -> CmdResult<Retu
                  VALUES (?1, 'ajuste', ?2, ?3, ?4, 'Devolución #' || ?5)",
                 params![
                     pid,
-                    item.qty as i64,
+                    item.qty,
                     qty_before,
-                    qty_before + item.qty as i64,
+                    qty_before + item.qty,
                     return_id,
                 ],
             )
